Change Product.Price type from double to decimal

+++ Grocery.Core/Models/Product.cs
@@ -8,13 +8,13 @@ namespace Grocery.Core.Models
         public int stock;
         public DateOnly ShelfLife { get; set; }
         
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         
         public Category? Category { get; set; }
 
-        public Product(int id, string name, int stock, double price) : this(id, name, stock, default, price) { }
+        public Product(int id, string name, int stock, decimal price) : this(id, name, stock, default, price) { }
 
-        public Product(int id, string name, int stock, DateOnly shelfLife, double price) : base(id, name)
+        public Product(int id, string name, int stock, DateOnly shelfLife, decimal price) : base(id, name)
         {
             Stock = stock;
             ShelfLife = shelfLife;
